Restructure the output generation part
Use webpack to generate the final browser app code and rewrote the
typescript player code in a separate file which is packed together with
the rest of the html code in a single js file.

Separated building the frontend and the golang code in the Makefile so
npm and webpack will not be invoked every time the go code will be
changed.

Changed the typescript writer code to use this new structure.

Other small changes here and there (will do a better job at splitting
the commits next time)