Use fresh and custom data when running webpack-dev-server
When running the webpack dev server, use the same data generated in the
current typescript in the current folder. This way it is easier to test
on some relevant data that can be generated every time a particular
feature is developed for the frontend.

The saved session data is extracted from typescript.html file in a
separate javascript file which is included by index.html which is then
served by webkac-dev-server